fix: address agent history sidebar review comments
