+++ apps/web/src/app-shell/SidebarPeekShell.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import React from "react";
 
 import { cn } from "@/shared/lib/utils";
+import { useSidebarPeekVisibility } from "@/app-shell/use-sidebar-peek-visibility";
 
 const SIDEBAR_PEEK_HIT_AREA_PX = 5;
-const SIDEBAR_PEEK_CLOSE_DELAY_MS = 160;
 
 interface SidebarPeekShellProps {
   side: "left" | "right";
@@ -31,76 +31,14 @@ export function SidebarPeekShell({
   widthPx,
   children,
 }: SidebarPeekShellProps) {
-  const triggerRef = useRef<HTMLDivElement>(null);
-  const panelRef = useRef<HTMLDivElement>(null);
-  const closeTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const [isVisible, setIsVisible] = useState(false);
-
-  const clearCloseTimer = useCallback(() => {
-    if (closeTimerRef.current) {
-      clearTimeout(closeTimerRef.current);
-      closeTimerRef.current = null;
-    }
-  }, []);
-
-  const showPeek = useCallback(() => {
-    clearCloseTimer();
-    setIsVisible(true);
-  }, [clearCloseTimer]);
-
-  const scheduleHide = useCallback(() => {
-    clearCloseTimer();
-    closeTimerRef.current = setTimeout(() => {
-      if (
-        triggerRef.current?.matches(":hover") ||
-        panelRef.current?.matches(":hover") ||
-        document.querySelector(SIDEBAR_PEEK_KEEP_OPEN_SELECTOR)
-      ) {
-        closeTimerRef.current = null;
-        return;
-      }
-
-      setIsVisible(false);
-      closeTimerRef.current = null;
-    }, SIDEBAR_PEEK_CLOSE_DELAY_MS);
-  }, [clearCloseTimer]);
-
-  const handlePointerLeave = useCallback(
-    (relatedTarget: EventTarget | null) => {
-      if (isSidebarPeekKeepOpenTarget(relatedTarget)) {
-        clearCloseTimer();
-        return;
-      }
-      scheduleHide();
-    },
-    [clearCloseTimer, scheduleHide],
-  );
-
-  useEffect(() => {
-    if (!isVisible) {
-      return;
-    }
-
-    const handlePointerOver = (event: PointerEvent) => {
-      const target = event.target;
-      if (
-        isNodeInsideRef(target, triggerRef) ||
-        isNodeInsideRef(target, panelRef) ||
-        isSidebarPeekKeepOpenTarget(target)
-      ) {
-        clearCloseTimer();
-        return;
-      }
-      scheduleHide();
-    };
-
-    document.addEventListener("pointerover", handlePointerOver, true);
-    return () => {
-      document.removeEventListener("pointerover", handlePointerOver, true);
-    };
-  }, [clearCloseTimer, isVisible, scheduleHide]);
-
-  useEffect(() => clearCloseTimer, [clearCloseTimer]);
+  const {
+    handleFocusLeave,
+    handlePointerLeave,
+    isVisible,
+    panelRef,
+    showPeek,
+    triggerRef,
+  } = useSidebarPeekVisibility();
 
   if (!collapsed) {
     return <div className="h-full w-full min-w-0">{children}</div>;
@@ -110,7 +48,7 @@ export function SidebarPeekShell({
   const edgeClassName = isLeft ? "left-0" : "right-0";
 
   return (
-    <>
+    <div className="contents" onFocusCapture={showPeek} onBlurCapture={handleFocusLeave}>
       <div
         ref={triggerRef}
         aria-hidden="true"
@@ -143,46 +81,11 @@ export function SidebarPeekShell({
           width:
             widthPx == null ? "min(360px, calc(100vw - 48px))" : `${widthPx}px`,
         }}
-        onFocusCapture={showPeek}
         onPointerEnter={showPeek}
         onPointerLeave={(event) => handlePointerLeave(event.relatedTarget)}
       >
         {children}
       </div>
-    </>
-  );
-}
-
-const SIDEBAR_PEEK_KEEP_OPEN_SELECTOR = [
-  "[data-workspace-popover-surface='true']:hover",
-  "[data-radix-popper-content-wrapper]:hover",
-  "[data-slot='popover-content']:hover",
-  "[data-slot='hover-card-content']:hover",
-  "[data-slot='tooltip-content']:hover",
-  "[data-slot='dropdown-menu-content']:hover",
-  "[data-slot='dropdown-menu-sub-content']:hover",
-].join(", ");
-
-const SIDEBAR_PEEK_KEEP_OPEN_TARGET_SELECTOR = [
-  "[data-workspace-popover-surface='true']",
-  "[data-radix-popper-content-wrapper]",
-  "[data-slot='popover-content']",
-  "[data-slot='hover-card-content']",
-  "[data-slot='tooltip-content']",
-  "[data-slot='dropdown-menu-content']",
-  "[data-slot='dropdown-menu-sub-content']",
-].join(", ");
-
-function isNodeInsideRef(
-  target: EventTarget | null,
-  ref: React.RefObject<HTMLElement | null>,
-) {
-  return target instanceof Node && ref.current?.contains(target);
-}
-
-function isSidebarPeekKeepOpenTarget(target: EventTarget | null) {
-  return (
-    target instanceof Element &&
-    Boolean(target.closest(SIDEBAR_PEEK_KEEP_OPEN_TARGET_SELECTOR))
+    </div>
   );
 }
